+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+ - Dragonfly add triplet
  - Dragonfly sources
  - Dragonfly sinks
  - Select nodes into dragonfly